refactor: extract CheckpointTimeline and SessionEventRow components
Decompose checkpoint and session-event rendering into dedicated components
to reduce coupling in the host views:

- CheckpointTimeline.vue: owns timeline layout, expand/collapse state,
  scroll-to-focus watcher, and all cp-* CSS. Accepts checkpoints +
  v-model focusNumber props.

- SessionEventRow.vue: renders a single session event (compaction with
  checkpoint pill, or severity-based regular event). Owns compaction
  detection, navigation injection, and all event/pill CSS.

OverviewTab.vue: -231 lines (616 -> 385). Now just drops in
<CheckpointTimeline> with 3 prop bindings.

ChatViewMode.vue: -128 lines. Session event loop replaced by
<SessionEventRow v-for>. Removed SessionEventSeverity/TurnSessionEvent
imports, severityClass/severityIcon/isCompactionEvent helpers,
useCheckpointNavigation import, and 80 lines of event CSS.
