+++ apps/desktop/src/components/conversation/ChatViewMode.vue
@@ -2,8 +2,6 @@
 import type {
   AttributedMessage,
   ConversationTurn,
-  SessionEventSeverity,
-  TurnSessionEvent,
   TurnToolCall,
 } from "@tracepilot/types";
 import { getToolArgs, toolArgString } from "@tracepilot/types";
@@ -25,7 +23,7 @@ import {
 } from "@tracepilot/ui";
 import { computed, nextTick, onMounted, onUnmounted, ref } from "vue";
 import { useRoute } from "vue-router";
-import { useCheckpointNavigation } from "@/composables/useCheckpointNavigation";
+import SessionEventRow from "@/components/conversation/SessionEventRow.vue";
 import { useCrossTurnSubagents } from "@/composables/useCrossTurnSubagents";
 import { useSubagentPanel } from "@/composables/useSubagentPanel";
 import { useToolResultLoader } from "@/composables/useToolResultLoader";
@@ -53,7 +51,6 @@ const preferences = usePreferencesStore();
 const route = useRoute();
 const turns = computed(() => store.turns);
 const renderMd = computed(() => preferences.isFeatureEnabled("renderMarkdown"));
-const navigateToCheckpoint = useCheckpointNavigation();
 
 // ─── Cross-turn subagent data ─────────────────────────────────────
 
@@ -188,22 +185,6 @@ const toolCallTurnIndex = computed(() => {
   }
   return map;
 });
-
-function severityClass(severity: SessionEventSeverity): string {
-  if (severity === "error") return "error";
-  if (severity === "warning") return "warning";
-  return "info";
-}
-
-function severityIcon(severity: SessionEventSeverity): string {
-  if (severity === "error") return "⚠️";
-  if (severity === "warning") return "⚠️";
-  return "ℹ️";
-}
-
-function isCompactionEvent(evt: TurnSessionEvent): boolean {
-  return evt.eventType === "session.compaction_complete";
-}
 
 // ─── Progressive disclosure ───────────────────────────────────────
 
@@ -459,41 +440,11 @@ defineExpose({ revealEvent });
               </div>
 
               <!-- Session events -->
-              <template v-for="evt in (turn.sessionEvents ?? [])" :key="evt.timestamp">
-                <!-- Compaction event with checkpoint pill -->
-                <div
-                  v-if="isCompactionEvent(evt)"
-                  class="cv-session-event cv-compaction"
-                >
-                  <span class="cv-session-event-icon">🗜️</span>
-                  <button
-                    v-if="evt.checkpointNumber != null"
-                    class="cv-checkpoint-pill"
-                    :title="`View Checkpoint #${evt.checkpointNumber} in Overview tab`"
-                    @click="navigateToCheckpoint(evt.checkpointNumber!)"
-                  >
-                    📋 Checkpoint #{{ evt.checkpointNumber }}
-                  </button>
-                  <span v-else class="cv-session-event-type">compaction</span>
-                  <span class="cv-session-event-summary">{{ evt.summary }}</span>
-                  <span v-if="evt.timestamp" class="cv-session-event-time">
-                    {{ formatTime(evt.timestamp) }}
-                  </span>
-                </div>
-
-                <!-- Regular session event -->
-                <div
-                  v-else
-                  :class="['cv-session-event', severityClass(evt.severity)]"
-                >
-                  <span class="cv-session-event-icon">{{ severityIcon(evt.severity) }}</span>
-                  <span class="cv-session-event-type">{{ evt.eventType }}</span>
-                  <span class="cv-session-event-summary">{{ evt.summary }}</span>
-                  <span v-if="evt.timestamp" class="cv-session-event-time">
-                    {{ formatTime(evt.timestamp) }}
-                  </span>
-                </div>
-              </template>
+              <SessionEventRow
+                v-for="evt in (turn.sessionEvents ?? [])"
+                :key="evt.timestamp"
+                :event="evt"
+              />
 
               <!-- User message anchor -->
               <div
@@ -781,86 +732,6 @@ defineExpose({ revealEvent });
   color: var(--text-placeholder, #6e7681);
   white-space: nowrap;
   user-select: none;
-}
-
-/* ─── Session events ───────────────────────────────────────────── */
-
-.cv-session-event {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  padding: 6px 12px;
-  border-radius: var(--radius-md, 8px);
-  font-size: 12px;
-  margin: 4px 0;
-}
-
-.cv-session-event.info {
-  background: var(--neutral-subtle, rgba(110, 118, 129, 0.1));
-  color: var(--text-secondary, #8b949e);
-}
-
-.cv-session-event.warning {
-  background: var(--warning-subtle, rgba(210, 153, 34, 0.1));
-  color: var(--warning-fg, #d29922);
-}
-
-.cv-session-event.error {
-  background: var(--danger-subtle, rgba(248, 81, 73, 0.1));
-  color: var(--danger-fg, #f85149);
-}
-
-.cv-session-event-icon {
-  flex-shrink: 0;
-}
-
-.cv-session-event-type {
-  font-family: "JetBrains Mono", monospace;
-  font-size: 11px;
-  opacity: 0.7;
-}
-
-.cv-session-event-summary {
-  flex: 1;
-  min-width: 0;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-}
-
-.cv-session-event-time {
-  flex-shrink: 0;
-  font-size: 11px;
-  opacity: 0.6;
-}
-
-/* ─── Compaction event ────────────────────────────────────────────── */
-
-.cv-compaction {
-  background: var(--accent-subtle, rgba(56, 139, 253, 0.08));
-  color: var(--accent-fg, #58a6ff);
-  border-left: 3px solid var(--accent-emphasis, #1f6feb);
-}
-
-.cv-checkpoint-pill {
-  display: inline-flex;
-  align-items: center;
-  gap: 4px;
-  padding: 1px 8px;
-  border: 1px solid var(--accent-muted, rgba(56, 139, 253, 0.4));
-  border-radius: 12px;
-  background: var(--accent-subtle, rgba(56, 139, 253, 0.1));
-  color: var(--accent-fg, #58a6ff);
-  font-size: 11px;
-  font-weight: 600;
-  cursor: pointer;
-  white-space: nowrap;
-  transition: background 0.15s, border-color 0.15s;
-}
-
-.cv-checkpoint-pill:hover {
-  background: var(--accent-muted, rgba(56, 139, 253, 0.25));
-  border-color: var(--accent-fg, #58a6ff);
 }
 
 /* ─── User message anchor ──────────────────────────────────────── */
